# Conflicts:
#	apps/server/src/modules/group/controller/api-test/group.api.spec.ts
#	apps/server/src/modules/group/controller/dto/request/class-sort-params.ts
#	apps/server/src/modules/group/controller/mapper/group-response.mapper.ts
#	apps/server/src/modules/oauth-provider/api/oauth-provider.login-flow.uc.spec.ts
#	apps/server/src/modules/oauth-provider/domain/service/id-token.service.spec.ts
#	apps/server/src/modules/oauth-provider/domain/service/oauth-provider.login-flow.service.spec.ts
#	apps/server/src/modules/oauth-provider/domain/service/oauth-provider.login-flow.service.ts
#	apps/server/src/modules/oauth-provider/oauth-provider-api.module.ts
#	apps/server/src/modules/oauth-provider/oauth-provider.module.ts
#	apps/server/src/modules/server/server.module.ts
#	apps/server/src/modules/tldraw/tldraw-api-test.module.ts
#	apps/server/src/modules/user-login-migration/index.ts
#	config/test.json

+++ apps/server/src/modules/learnroom/service/dashboard.service.ts
@@ -1,39 +1,71 @@
+import { MikroORM, UseRequestContext } from '@mikro-orm/core';
+import {
+	DataDeletedEvent,
+	DataDeletionDomainOperationLoggable,
+	DeletionService,
+	DomainDeletionReport,
+	DomainDeletionReportBuilder,
+	DomainName,
+	DomainOperationReportBuilder,
+	OperationType,
+	StatusModel,
+	UserDeletedEvent,
+} from '@modules/deletion';
 import { Inject, Injectable } from '@nestjs/common';
-import { DataDeletionDomainOperationLoggable } from '@shared/common/loggable';
-import { DomainModel, EntityId, StatusModel } from '@shared/domain/types';
-import { IDashboardRepo, DashboardElementRepo } from '@shared/repo';
+import { EventBus, EventsHandler, IEventHandler } from '@nestjs/cqrs';
+import { EntityId } from '@shared/domain/types';
+import { DashboardElementRepo, IDashboardRepo } from '@shared/repo';
 import { Logger } from '@src/core/logger';
 
 @Injectable()
-export class DashboardService {
+@EventsHandler(UserDeletedEvent)
+export class DashboardService implements DeletionService, IEventHandler<UserDeletedEvent> {
 	constructor(
 		@Inject('DASHBOARD_REPO') private readonly dashboardRepo: IDashboardRepo,
 		private readonly dashboardElementRepo: DashboardElementRepo,
-		private readonly logger: Logger
+		private readonly logger: Logger,
+		private readonly eventBus: EventBus,
+		private readonly orm: MikroORM
 	) {
 		this.logger.setContext(DashboardService.name);
 	}
 
-	async deleteDashboardByUserId(userId: EntityId): Promise<number> {
+	@UseRequestContext()
+	public async handle({ deletionRequestId, targetRefId }: UserDeletedEvent): Promise<void> {
+		const dataDeleted = await this.deleteUserData(targetRefId);
+		await this.eventBus.publish(new DataDeletedEvent(deletionRequestId, dataDeleted));
+	}
+
+	public async deleteUserData(userId: EntityId): Promise<DomainDeletionReport> {
 		this.logger.info(
 			new DataDeletionDomainOperationLoggable(
 				'Deleting user data from Dashboard',
-				DomainModel.DASHBOARD,
+				DomainName.DASHBOARD,
 				userId,
 				StatusModel.PENDING
 			)
 		);
-		const usersDashboard = await this.dashboardRepo.getUsersDashboard(userId);
-		await this.dashboardElementRepo.deleteByDashboardId(usersDashboard.id);
-		const result = await this.dashboardRepo.deleteDashboardByUserId(userId);
+		let deletedDashboard = 0;
+		const refs: string[] = [];
+		const usersDashboard = await this.dashboardRepo.getUsersDashboardIfExist(userId);
+		if (usersDashboard !== null) {
+			await this.dashboardElementRepo.deleteByDashboardId(usersDashboard.id);
+			deletedDashboard = await this.dashboardRepo.deleteDashboardByUserId(userId);
+			refs.push(usersDashboard.id);
+		}
+
+		const result = DomainDeletionReportBuilder.build(DomainName.DASHBOARD, [
+			DomainOperationReportBuilder.build(OperationType.DELETE, deletedDashboard, refs),
+		]);
+
 		this.logger.info(
 			new DataDeletionDomainOperationLoggable(
 				'Successfully deleted user data from Dashboard',
-				DomainModel.DASHBOARD,
+				DomainName.DASHBOARD,
 				userId,
 				StatusModel.FINISHED,
 				0,
-				result
+				deletedDashboard
 			)
 		);
